requirements.txt
cherrypy
pyttsx3
requests
pillow